Refactor: Andar Bahar game components and styles; introduce AndarBahar1 and AndarBahar2 variants
