+++ src/component/game/AndarBahar1/style/BetSection.module.css
