+++ src/component/game/AndarBahar2/style/BetSection.module.css
@@ -1,0 +1,31 @@
+.betSection {
+  height: 100%;
+  width: 100%;
+  display: flex;
+  justify-content: center;
+  flex-direction: column;
+  align-items: center;
+  gap: 10px;
+  padding: 10px 5px;
+}
+
+.betCard {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  gap: 5px;
+}
+.name {
+  font-size: calc(1.2 * var(--text-lg));
+  font-weight: 700;
+  color: var(--color-text-primary);
+}
+
+.betPlacing {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 10px;
+  width: 100%;
+}
