+++ src/component/game/AndarBahar2/style/CardSection.module.css
@@ -1,0 +1,6 @@
+.cardSection {
+  display: flex;
+  justify-content: center;
+  justify-items: center;
+  width: 100%;
+}
