+++ src/component/game/common/layout/jsx/BetSection.jsx
@@ -1,13 +1,15 @@
 import React from "react";
 import styles from "../style/BetSection.module.css";
 import { BetSection as Lucky7BBetSection } from "../../../lucky7B/jsx/BetSection";
-import { BetSection as AndarBaharBetSection } from "../../../AndarBahar/jsx/BetSection";
+import { BetSection as AndarBahar2BetSection } from "../../../AndarBahar2/jsx/BetSection";
+import { BetSection as AndarBahar1BetSection } from "../../../AndarBahar1/jsx/BetSection";
 import { BetSection as TeenPattiT20BetSection } from "../../../teenPattiT20/jsx/BetSection";
 
 export const BetSection = ({ game, onClick }) => {
   const betSectionMap = {
     lucky7B: Lucky7BBetSection,
-    andarBahar: AndarBaharBetSection,
+    andarBahar2: AndarBahar2BetSection,
+    andarBahar1: AndarBahar1BetSection,
     teenPattiT20: TeenPattiT20BetSection,
   };
