+++ src/component/game/common/layout/jsx/GameInterface.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styles from "../style/GameInterface.module.css";
 import { CardSection as Lucky7BCardSection } from "../../../lucky7B/jsx/CardSection";
-import { CardSection as AndarBaharCardSection } from "../../../AndarBahar/jsx/CardSection";
+import { CardSection as AndarBaharCardSection } from "../../../AndarBahar2/jsx/CardSection";
 import { CardSection as TeenPattiT20CardSection } from "../../../teenPattiT20/jsx/CardSection";
 
 export const GameInterface = ({ game, roundId = "A85745846" }) => {
@@ -13,7 +13,8 @@ export const GameInterface = ({ game, roundId = "A85745846" }) => {
 
   const gameNameMap = {
     lucky7B: "Lucky 7B",
-    andarBahar: "Andar Bahar",
+    andarBahar1: "Andar Bahar 1",
+    andarBahar2: "Andar Bahar 2",
     teenPattiT20: "Teen Patti T20",
   };
