+++ src/pages/jsx/Routing.jsx
@@ -40,9 +40,9 @@ export const Routing = () => {
           <Route
             path="/game"
             element={
-              <ProtectedRoutes allowedRoles={["player", "agent"]}>
+              // <ProtectedRoutes allowedRoles={["player", "agent"]}>
                 <Game />
-              </ProtectedRoutes>
+              // </ProtectedRoutes>
             }
           />
         </Routes>
